Conflicts:
	CMakeLists.txt
	include/Core/Core.h
	include/Data/Edge.h
	include/Data/Graph.h
	include/Data/GraphLayout.h
	include/Data/GraphSpanningTree.h
	include/Data/MetaType.h
	include/Data/Node.h
	include/Data/Type.h
	include/Importer/ImportInfoHandler.h
	include/Importer/ImportInfoHandlerEmpty.h
	include/Layout/FRAlgorithm.h
	include/Layout/RestrictionsManager.h
	include/Layout/ShapeVisitor.h
	include/Layout/ShapeVisitor_Comparator.h
	include/Layout/ShapeVisitor_DistanceFromFocusGetter.h
	include/Layout/ShapeVisitor_RestrictedPositionGetter.h
	include/Manager/ImportInfoHandlerImpl.h
	include/Math/CameraMath.h
	include/Math/DjikstraAlg.h
	include/Model/DB.h
	include/Model/EdgeDAO.h
	include/Model/GraphDAO.h
	include/Model/GraphLayoutDAO.h
	include/Model/NodeDAO.h
	include/Network/Client.h
	include/Network/Server.h
	include/OsgQtBrowser/QGraphicsViewAdapter.h
	include/OsgQtBrowser/QWebViewImage.h
	include/QOSG/AdapterWidget.h
	include/QOSG/CheckBoxList.h
	include/QOSG/CoreWindow.h
	include/QOSG/OptionsWindow.h
	include/QOSG/ViewerQT.h
	include/Viewer/CameraManipulator.h
	include/Viewer/CoreGraph.h
	include/Viewer/PerlinNoiseTextureGenerator.h
	include/Viewer/PickHandler.h
	include/Viewer/ShapeVisitor_ManipulatorCreator.h
	src/Core/Core.cpp
	src/Data/Edge.cpp
	src/Data/Graph.cpp
	src/Data/GraphSpanningTree.cpp
	src/Data/Node.cpp
	src/Data/Type.cpp
	src/Layout/FRAlgorithm.cpp
	src/Layout/ShapeGetter.cpp
	src/Layout/ShapeVisitor_Comparator.cpp
	src/Layout/ShapeVisitor_RestrictedPositionGetter.cpp
	src/Layout/Shape_ConeSurface.cpp
	src/Layout/Shape_CylinderSurface.cpp
	src/Manager/ImportInfoHandlerImpl.cpp
	src/Manager/Manager.cpp
	src/Math/CameraMath.cpp
	src/Math/DjikstraAlg.cpp
	src/Math/GraphMetrics.cpp
	src/Model/EdgeDAO.cpp
	src/Model/GraphDAO.cpp
	src/Model/NodeDAO.cpp
	src/Network/Client.cpp
	src/Network/Helper.cpp
	src/Network/Server.cpp
	src/Noise/noiseutils.cpp
	src/OsgQtBrowser/QGraphicsViewAdapter.cpp
	src/QOSG/AdapterWidget.cpp
	src/QOSG/CheckBoxList.cpp
	src/QOSG/CoreWindow.cpp
	src/Viewer/CameraManipulator.cpp
	src/Viewer/CoreGraph.cpp
	src/Viewer/DataHelper.cpp
	src/Viewer/NodeGroup.cpp
	src/Viewer/PerlinNoiseTextureGenerator.cpp
	src/Viewer/RestrictionManipulatorsGroup.cpp
	src/Viewer/ShapeVisitor_ManipulatorCreator.cpp
	src/Viewer/ShapeVisitor_VisualizerCreator.cpp
	src/main.cpp

+++ include/Math/CameraMath.h
@@ -3,7 +3,6 @@
 #include <osg/Vec3d>
 #include <osg/ref_ptr>
 #include <QVector>
-#include <QLinkedList>
 #include "Data/Node.h"
 
 class CameraMath
@@ -85,7 +84,8 @@ public:
 	 * \returns
 	 * point coordinates
 	 */
-	static osg::Vec3d getPointOnVector(osg::Vec3d p1, osg::Vec3d p2, float distance);
+
+	static osg::Vec3d getPointOnVector(osg::Vec3d p1, osg::Vec3d p2, double distance);
 
 	/*!
 	 * \brief
